requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
